requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
powerline-status==2.7
PyMySQL==1.1.1
marker-pdf==1.2.3
paddlepaddle==2.6.2
paddleocr==2.9.1
python-dotenv==1.0.1
openai==1.59.7
django-redis==5.4.0
websockets==14.1
channels==4.2.0
daphne==4.1.2
pydub==0.25.1
loguru==0.7.2
volcengine-python-sdk==1.0.122
oss2==2.19.1